Changes in Kafka source API
