+++ engine/flink/kafka-util/src/main/scala/pl/touk/nussknacker/engine/kafka/KafkaSourceFactory.scala
@@ -3,31 +3,21 @@ package pl.touk.nussknacker.engine.kafka
 import javax.validation.constraints.NotBlank
 import org.apache.flink.api.common.serialization.DeserializationSchema
 import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.streaming.api.TimeCharacteristic
-import org.apache.flink.streaming.api.functions.source.SourceFunction
-import org.apache.flink.streaming.api.functions.{AssignerWithPeriodicWatermarks, AssignerWithPunctuatedWatermarks, TimestampAssigner}
-import org.apache.flink.streaming.api.scala.{DataStream, StreamExecutionEnvironment}
-import org.apache.flink.streaming.connectors.kafka.internals.KafkaDeserializationSchemaWrapper
-import org.apache.flink.streaming.connectors.kafka.{FlinkKafkaConsumer, KafkaDeserializationSchema}
-import org.apache.kafka.clients.consumer.ConsumerRecord
-import org.apache.kafka.clients.producer.ProducerRecord
+import org.apache.flink.streaming.api.functions.TimestampAssigner
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.NodeId
 import pl.touk.nussknacker.engine.api.editor.{DualEditor, DualEditorMode, SimpleEditor, SimpleEditorType}
-import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, Source, TestDataGenerator, TestDataParserProvider}
-import pl.touk.nussknacker.engine.api.test.{TestDataParser, TestDataSplit}
+import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, Source, TestDataGenerator}
+import pl.touk.nussknacker.engine.api.test.TestDataSplit
 import pl.touk.nussknacker.engine.api.{MetaData, MethodToInvoke, ParamName}
-import pl.touk.nussknacker.engine.flink.api.compat.ExplicitUidInOperatorsSupport
-import pl.touk.nussknacker.engine.flink.api.process.{FlinkCustomNodeContext, FlinkSource, FlinkSourceFactory}
+import pl.touk.nussknacker.engine.flink.api.process.FlinkSourceFactory
 import pl.touk.nussknacker.engine.kafka.BaseKafkaSourceFactory._
 import pl.touk.nussknacker.engine.kafka.serialization.{FixedKafkaDeserializationSchemaFactory, KafkaDeserializationSchemaFactory}
-
-import scala.collection.JavaConverters._
 
 /** <pre>
   * Wrapper for [[org.apache.flink.streaming.connectors.kafka.FlinkKafkaConsumer]]
   * Features:
   *   - fetch latest N records which can be later used to test process in UI
-  * Fetching data is defined in [[pl.touk.nussknacker.engine.kafka.BaseKafkaSourceFactory.KafkaSource]] which
+  * Fetching data is defined in [[pl.touk.nussknacker.engine.kafka.KafkaSource]] which
   * extends [[pl.touk.nussknacker.engine.api.process.TestDataGenerator]]. See [[pl.touk.nussknacker.engine.kafka.KafkaUtils#readLastMessages]]
   *   - reset Kafka's offset to latest value - `forceLatestRead` property, see [[pl.touk.nussknacker.engine.kafka.KafkaUtils#setOffsetToLatest]]
   *
@@ -41,16 +31,18 @@ class KafkaSourceFactory[T: TypeInformation](deserializationSchemaFactory: Kafka
                                              timestampAssigner: Option[TimestampAssigner[T]],
                                              testPrepareInfo: TestDataSplit,
                                              processObjectDependencies: ProcessObjectDependencies)
-  extends BaseKafkaSourceFactory(timestampAssigner, testPrepareInfo, processObjectDependencies) {
+  extends BaseKafkaSourceFactory(deserializationSchemaFactory, timestampAssigner, testPrepareInfo, processObjectDependencies) {
 
-  def this(schema: DeserializationSchema[T],
+  def this(deserializationSchema: DeserializationSchema[T],
            timestampAssigner: Option[TimestampAssigner[T]],
            testPrepareInfo: TestDataSplit,
            processObjectDependencies: ProcessObjectDependencies) =
-    this(FixedKafkaDeserializationSchemaFactory(new KafkaDeserializationSchemaWrapper(schema)),
+    this(
+      FixedKafkaDeserializationSchemaFactory(deserializationSchema),
       timestampAssigner,
       testPrepareInfo,
-      processObjectDependencies)
+      processObjectDependencies
+    )
 
   @MethodToInvoke
   def create(processMetaData: MetaData,
@@ -61,7 +53,7 @@ class KafkaSourceFactory[T: TypeInformation](deserializationSchemaFactory: Kafka
              @ParamName(`TopicParamName`) @NotBlank topic: String)
             (implicit nodeId: NodeId): Source[T] with TestDataGenerator = {
     val kafkaConfig = KafkaConfig.parseProcessObjectDependencies(processObjectDependencies)
-    createSource(List(topic), kafkaConfig, deserializationSchemaFactory, processMetaData, nodeId)
+    createSource(List(topic), kafkaConfig, processMetaData, nodeId)
   }
 }
 
@@ -74,15 +66,16 @@ class SingleTopicKafkaSourceFactory[T: TypeInformation](topic: String,
                                                         timestampAssigner: Option[TimestampAssigner[T]],
                                                         testPrepareInfo: TestDataSplit,
                                                         processObjectDependencies: ProcessObjectDependencies)
-  extends BaseKafkaSourceFactory(timestampAssigner, testPrepareInfo, processObjectDependencies) {
+  extends BaseKafkaSourceFactory(deserializationSchemaFactory, timestampAssigner, testPrepareInfo, processObjectDependencies) {
 
   def this(topic: String,
-           schema: DeserializationSchema[T],
+           deserializationSchema: DeserializationSchema[T],
            timestampAssigner: Option[TimestampAssigner[T]],
            testPrepareInfo: TestDataSplit,
            processObjectDependencies: ProcessObjectDependencies) =
-    this(topic, FixedKafkaDeserializationSchemaFactory(
-      new KafkaDeserializationSchemaWrapper(schema)),
+    this(
+      topic,
+      FixedKafkaDeserializationSchemaFactory(deserializationSchema),
       timestampAssigner,
       testPrepareInfo,
       processObjectDependencies
@@ -91,93 +84,23 @@ class SingleTopicKafkaSourceFactory[T: TypeInformation](topic: String,
   @MethodToInvoke
   def create(processMetaData: MetaData)(implicit nodeId: NodeId): Source[T] with TestDataGenerator = {
     val kafkaConfig = KafkaConfig.parseProcessObjectDependencies(processObjectDependencies)
-    createSource(List(topic), kafkaConfig, deserializationSchemaFactory, processMetaData, nodeId)
+    createSource(List(topic), kafkaConfig, processMetaData, nodeId)
   }
 }
 
-abstract class BaseKafkaSourceFactory[T: TypeInformation](val timestampAssigner: Option[TimestampAssigner[T]],
-                                                          protected val testPrepareInfo: TestDataSplit,
+abstract class BaseKafkaSourceFactory[T: TypeInformation](deserializationSchemaFactory: KafkaDeserializationSchemaFactory[T],
+                                                          timestampAssigner: Option[TimestampAssigner[T]],
+                                                          testPrepareInfo: TestDataSplit,
                                                           processObjectDependencies: ProcessObjectDependencies)
   extends FlinkSourceFactory[T] with Serializable {
 
-  @deprecated("Should be used version without process MetaData", "0.1.1")
-  protected def createSource(processMetaData: MetaData, topics: List[String], schema: KafkaDeserializationSchema[T]): KafkaSource =
-    createSource(topics, KafkaConfig.parseProcessObjectDependencies(processObjectDependencies), FixedKafkaDeserializationSchemaFactory(schema))
-
   // We currently not using processMetaData and nodeId but it is here in case if someone want to use e.g. some additional fields
   // in their own concrete implementation
-  protected def createSource(topics: List[String],
-                             kafkaConfig: KafkaConfig,
-                             deserializationSchemaFactory: KafkaDeserializationSchemaFactory[T],
-                             processMetaData: MetaData, nodeId: NodeId): KafkaSource =
-    createSource(topics, kafkaConfig, deserializationSchemaFactory)
+  protected def createSource(topics: List[String], kafkaConfig: KafkaConfig, processMetaData: MetaData, nodeId: NodeId): KafkaSource[T] =
+    createSource(topics, kafkaConfig)
 
-  protected def createSource(topics: List[String], kafkaConfig: KafkaConfig, deserializationSchemaFactory: KafkaDeserializationSchemaFactory[T]): KafkaSource = {
+  protected def createSource(topics: List[String], kafkaConfig: KafkaConfig): KafkaSource[T] = {
     val serializationSchema = deserializationSchemaFactory.create(topics, kafkaConfig)
-    new KafkaSource(topics = topics, kafkaConfig, serializationSchema, None, processObjectDependencies)
-  }
-
-  class KafkaSource(topics: List[String],
-                    kafkaConfig: KafkaConfig,
-                    schema: KafkaDeserializationSchema[T],
-                    recordFormatterOpt: Option[RecordFormatter],
-                    processObjectDependencies: ProcessObjectDependencies,
-                    overriddenConsumerGroup: Option[String] = None)
-    extends FlinkSource[T]
-      with Serializable
-      with TestDataParserProvider[T]
-      with TestDataGenerator with ExplicitUidInOperatorsSupport {
-
-    override def sourceStream(env: StreamExecutionEnvironment, flinkNodeContext: FlinkCustomNodeContext): DataStream[T] = {
-      val consumerGroupId = overriddenConsumerGroup.getOrElse(ConsumerGroupDeterminer(kafkaConfig).consumerGroup(flinkNodeContext))
-      env.setStreamTimeCharacteristic(if (timestampAssigner.isDefined) TimeCharacteristic.EventTime else TimeCharacteristic.IngestionTime)
-
-      val newStart = setUidToNodeIdIfNeed(flinkNodeContext,
-        env
-          .addSource[T](flinkSourceFunction(consumerGroupId))(typeInformation)
-          .name(s"${flinkNodeContext.metaData.id}-${flinkNodeContext.nodeId}-source"))
-
-      timestampAssigner.map {
-        case periodic: AssignerWithPeriodicWatermarks[T@unchecked] =>
-          newStart.assignTimestampsAndWatermarks(periodic)
-        case punctuated: AssignerWithPunctuatedWatermarks[T@unchecked] =>
-          newStart.assignTimestampsAndWatermarks(punctuated)
-      }.getOrElse(newStart)
-    }
-
-    def preparedTopics: List[String] = topics.map(KafkaUtils.prepareTopicName(_, processObjectDependencies))
-
-    protected val typeInformation: TypeInformation[T] = implicitly[TypeInformation[T]]
-
-    protected def flinkSourceFunction(consumerGroupId: String): SourceFunction[T] = {
-      preparedTopics.foreach(KafkaUtils.setToLatestOffsetIfNeeded(kafkaConfig, _, consumerGroupId))
-      createFlinkSource(consumerGroupId)
-    }
-
-    protected def createFlinkSource(consumerGroupId: String): FlinkKafkaConsumer[T] = {
-      new FlinkKafkaConsumer[T](preparedTopics.asJava, schema, KafkaUtils.toProperties(kafkaConfig, Some(consumerGroupId)))
-    }
-
-    override def generateTestData(size: Int): Array[Byte] = {
-      val listsFromAllTopics = preparedTopics.map(KafkaUtils.readLastMessages(_, size, kafkaConfig))
-      val merged = ListUtil.mergeListsFromTopics(listsFromAllTopics, size)
-      val formatted = recordFormatterOpt.map(formatter => merged.map(formatter.formatRecord)).getOrElse {
-        merged.map(_.value())
-      }
-      testPrepareInfo.joinData(formatted)
-    }
-
-    override def testDataParser: TestDataParser[T] = new TestDataParser[T] {
-      override def parseTestData(merged: Array[Byte]): List[T] =
-        testPrepareInfo.splitData(merged).map { formatted =>
-          val topic = preparedTopics.head
-          val record = recordFormatterOpt
-            .map(formatter => formatter.parseRecord(formatted))
-            .getOrElse(new ProducerRecord(topic, formatted))
-          schema.deserialize(new ConsumerRecord[Array[Byte], Array[Byte]](topic, -1, -1, record.key(), record.value()))
-        }
-    }
-
-    override def timestampAssignerForTest : Option[TimestampAssigner[T]] = BaseKafkaSourceFactory.this.timestampAssigner
+    new KafkaSource(topics = topics, kafkaConfig, serializationSchema, timestampAssigner, None, testPrepareInfo, processObjectDependencies)
   }
 }
