+++ engine/flink/kafka-util/src/main/scala/pl/touk/nussknacker/engine/kafka/serialization/KafkaDeserializationSchemaFactory.scala
@@ -1,6 +1,8 @@
 package pl.touk.nussknacker.engine.kafka.serialization
 
+import org.apache.flink.api.common.serialization.DeserializationSchema
 import org.apache.flink.streaming.connectors.kafka.KafkaDeserializationSchema
+import org.apache.flink.streaming.connectors.kafka.internals.KafkaDeserializationSchemaWrapper
 import pl.touk.nussknacker.engine.kafka.KafkaConfig
 
 /**
@@ -25,4 +27,9 @@ case class FixedKafkaDeserializationSchemaFactory[T](deserializationSchema: Kafk
 
   override def create(topics: List[String], kafkaConfig: KafkaConfig): KafkaDeserializationSchema[T] =
     deserializationSchema
+}
+
+object FixedKafkaDeserializationSchemaFactory {
+  def apply[T](deserializationSchema: DeserializationSchema[T]): FixedKafkaDeserializationSchemaFactory[T] =
+    new FixedKafkaDeserializationSchemaFactory(new KafkaDeserializationSchemaWrapper(deserializationSchema))
 }
